requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.0.2
django-allauth==65.9.0
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
openai==0.28.0
packaging==25.0
phonenumbers==9.0.6
propcache==0.3.1
psycopg2==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.0
